Add config, fix build

+++ src/tinyexr.hpp
@@ -2292,7 +2292,7 @@ static bool hufBuildDecTable(const long long *hcode,  // i : encoding table
         unsigned int *p = pl->p;
         pl->p = new unsigned int[pl->lit];
 
-        for (int i = 0; i < pl->lit - 1; ++i) pl->p[i] = p[i];
+        for (int i = 0; i < int(pl->lit) - 1; ++i) pl->p[i] = p[i];
 
         delete[] p;
       } else {
@@ -2548,7 +2548,7 @@ static bool hufDecode(const long long *hcode,  // i : encoding table
 
         int j;
 
-        for (j = 0; j < pl.lit; j++) {
+        for (j = 0; unsigned(j) < pl.lit; j++) {
           int l = hufLength(hcode[pl.p[j]]);
 
           while (lc < l && in < ie)  // get more bits
@@ -2570,7 +2570,7 @@ static bool hufDecode(const long long *hcode,  // i : encoding table
           }
         }
 
-        if (j == pl.lit) {
+        if (unsigned(j) == pl.lit) {
           return false;
           // invalidCode(); // Not found
         }
